Move GET /like/{id} to POST /item/{id}/activity/{verb}
