+++ view/js/main.js
@@ -669,7 +669,7 @@ function doActivityItem(ident, verb, un) {
 	unpause();
 	$('#like-rotator-' + ident.toString()).show();
 	verb = un ? 'un' + verb : verb;
-	$.get('like/' + ident.toString() + '?verb=' + verb, NavUpdate);
+	$.post('item/' + ident.toString() + '/activity/' + verb, NavUpdate);
 	liking = 1;
 	force_update = true;
 	update_item = ident.toString();
